Move BrushScene class to separate file
